consume me daddy
consume

+++ src/Items/Consumable.java
@@ -2,7 +2,7 @@ package Items;
 
 public interface Consumable
 {
-    public double consume();
+    public int consume();
     public int health();
     //public void dispose();
     //public default void refresh() { System.out.println("Refreshing"); }
